+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/dto/SysSystemAttributeMappingDto.java
@@ -24,6 +24,7 @@ public class SysSystemAttributeMappingDto extends AbstractDto implements Attribu
 	private boolean extendedAttribute = false;
 	private boolean entityAttribute = true;
 	private boolean confidentialAttribute = false;
+	private boolean passwordAttribute = false;
 	private boolean uid = false;
 	private String transformFromResourceScript;
 	private String transformToResourceScript;
@@ -196,5 +197,15 @@ public class SysSystemAttributeMappingDto extends AbstractDto implements Attribu
 	
 	public void setSendOnPasswordChange(boolean sendOnPasswordChange) {
 		this.sendOnPasswordChange = sendOnPasswordChange;
+	}
+
+	@Override
+	public boolean isPasswordAttribute() {
+		return passwordAttribute;
+	}
+
+	@Override
+	public void setPasswordAttribute(boolean passwordAttribute) {
+		this.passwordAttribute = passwordAttribute;
 	}
 }