+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/entity/SysSystemAttributeMapping.java
@@ -75,6 +75,10 @@ public class SysSystemAttributeMapping extends AbstractEntity {
 	@Audited
 	@Column(name = "confidential_attribute", nullable = false)
 	private boolean confidentialAttribute = false;
+	
+	@Audited
+	@Column(name = "password_attribute", nullable = false)
+	private boolean passwordAttribute = false;
 
 	@Audited
 	@Column(name = "uid", nullable = false)
@@ -250,5 +254,13 @@ public class SysSystemAttributeMapping extends AbstractEntity {
 
 	public void setCached(boolean cached) {
 		this.cached = cached;
+	}
+
+	public boolean isPasswordAttribute() {
+		return passwordAttribute;
+	}
+
+	public void setPasswordAttribute(boolean passwordAttribute) {
+		this.passwordAttribute = passwordAttribute;
 	}
 }