+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/DefaultSysProvisioningOperationService.java
@@ -316,6 +316,13 @@ public class DefaultSysProvisioningOperationService
 								((ConfidentialString) attribute.getValue()).getKey()));
 			} else if (attribute instanceof IcPasswordAttribute && attribute.getValue() == null) {
 				attributeCopy = new IcPasswordAttributeImpl(attribute.getName(), (GuardedString) null);
+			} else if (attribute.getValue() instanceof ConfidentialString) {
+				attributeCopy = new IcAttributeImpl(
+						attribute.getName(),
+						confidentialStorage.getGuardedString(
+						provisioningOperation.getId(),
+						SysProvisioningOperation.class,
+						((ConfidentialString) attribute.getValue()).getKey()));
 			} else {
 				attributeCopy = new IcAttributeImpl(attribute.getName(), attribute.getValue());
 			}